Added all ser/deserialization methods for new classes + added them in H.O.W. added singleLinkedList

+++ src/java/org/apache/hadoop/hbase/io/GetFamilies.java
@@ -2,11 +2,11 @@ package org.apache.hadoop.hbase.io;
 
 public class GetFamilies extends AbstractGet{
   
-  public GetFamilies(byte [] row, byte [] family, int versions, TimeRange tr){
+  public GetFamilies(byte [] row, byte [] family, byte versions, TimeRange tr){
     super.row = row;
     super.versions = versions;
     super.families = new Family[]{new Family(family)};
     super.tr = tr;
   }
-  
+
 }
